great work today! Time to work on the login and jwt auth
