+++ client/src/App.css
@@ -1,55 +1,1 @@
-.App {
-    width: 100vw;
-    height: 100vh;
-}
 
-body {
-    margin: 0;
-    padding: 0;
-}
-
-.Sidebar{
-    height:100%;
-    width:280px;
-    background-color: #2F4050;
-}
-
-.SidebarList {
-    height:auto;
-    width:100%;
-    padding: 0;
-    
-}
-
-.SidebarList .row {
-   width:100%;
-   height: 60px;
-   list-style-type: none; 
-   margin: 0%;
-   display: flex;
-   flex-direction: row;
-   color: #fff;
-   justify-content: center;
-   align-items: center;
-   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-
-}
-
-.SidebarList .row:hover {
-cursor: pointer;
-background-color: #293846;
-}
-
-.SidebarList .row #active{
-    background-color: #293846;
-}
-
-.row #icon {
-    flex: 30%;
-    display: grid;
-    place-items: center;
-}
-    
-.row #title {
-    flex: 70%;
-}
