--- client/src/Components/Sidebar.js
@@ -1,40 +1,0 @@
-import React from 'react'
-import '../App.css';
-import {SidebarData} from './SidebarData'
-
-
-function Sidebar() {
-
-    return (
-    <div className="Sidebar">
-       
-       <ul lassName="SidebarList"> 
-    {SidebarData.map((val, key)=> {
-     
-     return (
-         
-        <li key={key}
-        className="row" 
-        id={Window.location.pathname ==val.link ? "active" : ""}
-        onClick={()=> {
-            window.location.pathname = val.link
-            }}> 
-        
-         <div id="icon">{val.icon}</div> 
-         <div id="title">{val.title}</div>
-             
-         
-         </li>
-     );
-        
-    })}
-    
-    </ul>
-    
-    </div>       
-    
-    );   
-    
-}
-
-export default Sidebar;
